Refactoring template frame into reuseable component
