+++ frontend/signin.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { createRoot } from 'react-dom/client';
-import SignInSide from './sign-in-side/SignInSide';
+import SignInSide from './signin/SignInSide';
 
 document.body.style.margin = '0px';
 const container = document.getElementById('root');
